requirements.txt
magnet2torrent==1.2.1
requests==2.31.0
python-Levenshtein==0.25.0
termcolor==2.4.0